requirements.txt
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
openai>=1.3.0
click>=8.1.7
grpcio>=1.59.0
grpcio-tools>=1.59.0
fastapi>=0.104.0
uvicorn>=0.24.0
pytest>=7.4.0
pytest-asyncio>=1.3.0
aiohttp>=3.12.0
loguru>=0.7.2
pyyaml>=6.0.1
pillow>=10.0.0
pystray>=0.19.5
sqlite-vec>=0.1.6
sqlalchemy>=2.0.0
cryptography>=41.0.0
watchdog>=3.0.0

# Consolidated from legacy repos
opencv-python>=4.8.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
pynput>=1.7.6
mss>=9.0.1
numpy>=1.25.1
pyperclip>=1.9.0
requests>=2.32.3
pypresence>=4.3.0
pywin32>=306
psutil>=5.9.8
